Extracted IntegrationTest and introduced LobbyTests.
