+++ Whist.Server.Tests/LobbyTests.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Whist.Server.Tests
+{
+    public class LobbyTests : IntegrationTest
+    {
+        [Test]
+        public async Task LobbyTest()
+        {
+            Assert.Pass();
+        }
+    }
+}
